moved AuthArea server component inside navbar client component
